requirements.txt
streamlit
streamlit-cookies-manager
requests
beautifulsoup4
google-auth
google-auth-oauthlib
google-api-python-client
pandas
openpyxl